text_recognition_lambda/requirements.txt
boto3==1.26.10
pillow==9.3.0
pytesseract==0.3.9
